handling 404 error

+++ src/data_collection/coletor_logomarcas.php
@@ -1,38 +1,40 @@
-<?php  
+<?php
 /*
-> Sprint 1 - US 2
-> Danilo Gomes
-> Codigo para pegar as logomarcas das empresas que estão na bolsa de valores no site infomoney
-> Tem como saida as imagens das logomarcas das empresas que existem no site no formato .jpg, com o nome do 
-arquivo sendo o nome de pregão da empresa
-
-*/
-
+ * > Sprint 1 - US 2 > Danilo Gomes > Codigo para pegar as logomarcas das empresas que estão na bolsa de valores no site infomoney > Tem como saida as imagens das logomarcas das empresas que existem no site no formato .jpg, com o nome do arquivo sendo o nome de pregão da empresa
+ */
 $url = 'http://www.infomoney.com.br/mercados/empresas-bovespa';
-$html = file_get_contents('http://www.infomoney.com.br/mercados/empresas-bovespa');
- 
-$dom = new DOMDocument();
-$dom->loadHTML($html);
+$html = file_get_contents ( 'http://www.infomoney.com.br/mercados/empresas-bovespa' );
 
-$xpath = new DomXPath($dom);
-$dom = $xpath->query('//div[@class="por-setor ordem-alfa"]')->item(0);
+$dom = new DOMDocument ();
+$dom->loadHTML ( $html );
 
-$links = $dom->getElementsByTagName('a');
+$xpath = new DomXPath ( $dom );
+$dom = $xpath->query ( '//div[@class="por-setor ordem-alfa"]' )->item ( 0 );
 
-foreach ($links as $link){
-	echo $link->getAttribute('href').'--'.$link->getAttribute('title');
+$links = $dom->getElementsByTagName ( 'a' );
+
+foreach ( $links as $link ) {
+	echo $link->getAttribute ( 'href' ) . '--' . $link->getAttribute ( 'title' );
 	
-	$href = $link->getAttribute('href');
-	$html = file_get_contents('http://www.infomoney.com.br/' . $href);
+	$href = 'http://www.infomoney.com.br/' . $link->getAttribute ( 'href' );
 	
-	$dom = new DOMDocument();
-	$dom->loadHTML($html);
-	$img = $dom->getElementByID('imgLogoCompany');
-	
-	$img_filename = $img->getAttribute('title').'.jpg';
-	$img_src = $img->getAttribute('src');
-	if (!file_exists('icons')) {
-		mkdir('icons', 0777, true);
+	if (get_http_response_code ( $href ) != "404") {
+		$html = file_get_contents ( $href );
+		
+		$dom = new DOMDocument ();
+		$dom->loadHTML ( $html );
+		
+		$img = $dom->getElementByID ( 'imgLogoCompany' );
+		
+		$img_filename = $img->getAttribute ( 'title' ) . '.jpg';
+		if(!file_exists($img_filename)){
+			$img_src = $img->getAttribute ( 'src' );
+			if (! file_exists ( 'icons' )) {
+				mkdir ( 'icons', 0777, true );
+			}
+			file_put_contents ( "icons/" . $img_filename, file_get_contents ( 'http://www.infomoney.com.br/' . $img_src ) );
+		}
+	} else {
+		echo "error downloading: " . $href;
 	}
-	file_put_contents("icons/".$img_filename, file_get_contents('http://www.infomoney.com.br/' . $img_src));
 }
